test code coverage, refactored, added conjured support
notice that conjured does not support mixing with legendary items and/or stuff that increases in quality after a day has passed.

also note that it is deceptive in reporting code coverage, since the main() function is not tested (obviously). if you check coveragereport/index.html, it should be clear to see that every other method has every branch hit.